requirements.txt
Flask==2.3.3
waitress==2.1.2
qrcode==7.4.2
Pillow==10.2.0
xhtml2pdf==0.2.11
pandas==2.1.3
openpyxl==3.1.2
pyserial==3.5
requests==2.31.0
